Change get_angle() return type

+++ src/Futaba30xServo.hpp
@@ -19,7 +19,7 @@ public:
 	int move ( unsigned int deg, unsigned int time_10ms, unsigned char flag = 0x00 );
 	int torque_on ( unsigned char flag = 0x00 );
 	int torque_off ( unsigned char flag = 0x00 );
-	int get_angle();
+	short get_angle();
 	~Futaba30x ()
 	{
     close(fd);
